app_bot/requirements.txt
fastapi
uvicorn
python-multipart
httpx
python-dotenv
pydantic
email-validator
